+++ v2/migrate/v1/types.go
@@ -1,4 +1,4 @@
-package v0
+package v1
 
 import (
 	"time"
